Fix how Intake handles multiValued entries. Initial patch to IntegerField provided by Derek Gaasch.
git-svn-id: https://svn.apache.org/repos/asf/jakarta/turbine/core/branches/TURBINE_2_3_BRANCH@221831 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/turbine/services/intake/model/IntegerField.java
@@ -23,6 +23,8 @@ import org.apache.turbine.services.intake.validator.IntegerValidator;
 import org.apache.turbine.services.intake.xmlmodel.XmlField;
 
 /**
+ * Processor for int fields.
+ *
  * @version $Id$
  */
 public class IntegerField
@@ -112,11 +114,12 @@ public class IntegerField
         if (isMultiValued)
         {
             String[] inputs = parser.getStrings(getKey());
-            Integer[] values = new Integer[inputs.length];
+            int[] values = new int[inputs.length];
             for (int i = 0; i < inputs.length; i++)
             {
                 values[i] = StringUtils.isNotEmpty(inputs[i])
-                        ? new Integer(inputs[i]) : (Integer) getEmptyValue();
+                        ? new Integer(inputs[i]).intValue() 
+                        : ((Integer) getEmptyValue()).intValue();
             }
             setTestValue(values);
         }
@@ -126,4 +129,5 @@ public class IntegerField
             setTestValue(StringUtils.isNotEmpty(val) ? new Integer(val) : (Integer) getEmptyValue());
         }
     }
+    
 }
